# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A secure file storage REST API with S3 and CloudFront",
        "title": "File Vault API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mohamed8eo/file-vault",
            "email": "support@filevault.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http", "https"],
    "paths": {
...
        "/auth/google": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Initiate Google OAuth",
                "description": "Redirect to Google for authentication",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "cli",
                    "in": "query",
                    "type": "string",
                    "description": "Client type (web/mobile)"
                }],
                "responses": {
                    "302": {"description": "Redirects to Google"}
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Google OAuth callback",
                "description": "Handle Google OAuth callback and issue tokens",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "code",
                    "in": "query",
                    "type": "string",
                    "required": true,
                    "description": "Authorization code"
                }, {
                    "name": "state",
                    "in": "query",
                    "type": "string",
                    "description": "State parameter"
                }],
                "responses": {
                    "200": {"description": "OAuth successful"}
                }
            }
        },
        "/auth/github": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Initiate GitHub OAuth",
                "description": "Redirect to GitHub for authentication",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "cli",
                    "in": "query",
                    "type": "string",
                    "description": "Client type (web/mobile)"
                }],
                "responses": {
                    "302": {"description": "Redirects to GitHub"}
                }
            }
        },
        "/auth/github/callback": {
            "get": {
                "tags": ["Authentication"],
                "summary": "GitHub OAuth callback",
                "description": "Handle GitHub OAuth callback and issue tokens",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "code",
                    "in": "query",
                    "type": "string",
                    "required": true,
                    "description": "Authorization code"
                }, {
                    "name": "state",
                    "in": "query",
                    "type": "string",
                    "description": "State parameter"
                }],
                "responses": {
                    "200": {"description": "OAuth successful"}
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "Type 'Bearer' followed by a space and JWT token."
        }
    }
}